requirements.txt
# Core Libraries
moviepy


# Math and Data Manipulation Libraries
numpy
scipy
matplotlib
opencv
opencv-contrib-python
# To Experiment with:

# Graphical libraries
pygfx
jupyter-rfb
pycairo


# These are the libraries for manipulating audio
pydub


# These are the libraries for generating/capturing video
mss
dxcam

imageio
imageio-ffmpeg